made a figure to plot on and added a title, now need some labels...
